Experimental graphics library, 'aqwin'
It is included here to power my future graphics demos. It may evolve to
be a proper library in the future. I've factored it out so that the
system-specific parts are in separate files. Currently it only builds
on top of SDL2.

It's a little unique in that it uses text strings to configure the
window instead of traditional enums and bit masks. We'll see if that's
a good thing or a bad thing :)